requirements_enhanced.txt
# Enhanced Medical Image Analysis Requirements
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0

# Web Interface
gradio>=3.40.0

# Data Processing
tqdm>=4.65.0
scikit-learn>=1.3.0
reportlab>=4.0.0

# Optional: For better performance
# tensorrt  # NVIDIA GPU acceleration (optional)
# onnx      # Model optimization (optional)

